corfu-insert: Do not insert first candidate if no candidate is selected (Fix #152)
This special behavior of corfu-insert occurred with corfu-preselect-first set to
nil. The behavior was a remnant from before the introduction of the
corfu-preselect-first customization option. Since corfu-preselect-first has been
introduced a long time ago and I haven't heard complaints about it, let's now
remove the special behavior of corfu-insert. This change simplifies the
corfu--insert function slightly and makes the behavior of corfu-insert more
consistent with other Corfu behavior. If the new behavior conflicts with your
workflow, please let me know and we can look into possible workarounds. Feedback
on this issue is welcome.