corfu--in-region, corfu--auto-complete: Fix gud-gdb completion (Fix #157)
Copy the start marker returned by the `completion-at-point-function' only if is
not already a marker. This ensures that markers created by the
`completion-at-point-function' (and in particular its marker type) are kept
intact. The `gud-gdb-completion-at-point' has a hack which uses an insert-before
marker to hide some side effects of the completion process. It is unclear to me
if always copying the markers is disallowed by the `completion-at-point' API. I
believe the approach used by Corfu was not wrong. But this means then that
`gud-gdb-completion-at-point' must be improved such that it doesn't rely on the
insert-before hack anymore. The FIXME in `gud-gdb-completion-at-point` also
hints at that.