Create candidate preview before showing popup
The overlay previewing the current completion candidate may shift
subsequent text, including overlays used for popups due to running in
the terminal where child frames are unavailable. E.g.

     f|oobar long line that wr\
    aps +--------+
    |foo     |
    |>foobar |
    +--------+

where "|" is the cursor, and "\" indicates line wrapping and ">" the
current candidate. Therefore it is necessary to create the candidate
preview before positioning any completion popup overlays.

This commit also moves the call to redisplay into the child frame
specific code, since it is redundant when overlays are used to display
the popup, instead only causing flickering in the terminal.