Do not quit on empty input
When entering a trigger character, e.g., a dot, the popup will show even on
empty input. This change prevents flicker for the sequence of inputs: "obj. ->
obj.fun -> obj.". The downside of the change is backspacing the popup doesn't
work anymore in all cases - it depends on the Capf.