refactor: favor /*-let\*/ macros
I am noticing the following in my use of 31.0.50 version of Emacs (built
from HEAD of `master` branch):

> Warning: ‘if-let’ is an obsolete macro (as of 31.1); use ‘if-let*’
> instead.

And:

> Warning: ‘when-let’ is an obsolete macro (as of 31.1); use ‘when-let*’
> or ‘and-let*’ instead.

My understanding is that corfu requires 29.1 of Emacs.  Reviewing the
29.1 version of Emacs, the `if-let*` and `when-let*` macro is defined:

```sh
❯ cd ~/git/emacs; git checkout emacs-29.1; rg "defmacro (if|when)-let\*"
HEAD is now at a9b28224af0 ; Last-minute changes befor releasing 29.1
lisp/subr.el
2479:(defmacro if-let* (varlist then &rest else)
2493:(defmacro when-let* (varlist &rest body)
```

Thus, we should be able to swtich without breaking functionality from
the `if-let` to `if-let*` and likewise for `when-let`.